♻️ git login 기능 반환 형식 변경
- git login 후 로그인 성공 페이지로 토큰 담아 리다이렉션하도록 변경
- GithubLoginController 패키지 위치 이동
